+++ app/mailers/report_mailer.rb
@@ -34,9 +34,19 @@ class ReportMailer < ApplicationMailer
     @user = params[:user]
     raise_unless_active_user
 
-    view_mail(ENV["NOTIFY_VIEW_TEMPLATE"],
-      to: @user.email,
-      subject: t("mailer.report.approved.subject", application_name: t("app.title")))
+    @role = if @user.organisation == @report_presenter.organisation
+      :delivery_partner
+    elsif @user.organisation.service_owner?
+      :service_owner
+    end
+
+    if @role.present?
+      view_mail(ENV["NOTIFY_VIEW_TEMPLATE"],
+        to: @user.email,
+        subject: t("mailer.report.approved.#{@role}.subject", application_name: t("app.title")))
+    else
+      raise ArgumentError, "User must either be a service owner or belong to the organisation making the report"
+    end
   end
 
   def awaiting_changes