+++ app/models/organisation.rb
@@ -10,6 +10,7 @@ class Organisation < ApplicationRecord
   enum role: {
     delivery_partner: 0,
     matched_effort_provider: 1,
+    external_income_provider: 2,
     service_owner: 99,
   }
 
@@ -20,16 +21,18 @@ class Organisation < ApplicationRecord
   validates :iati_reference,
     uniqueness: {case_sensitive: false},
     presence: true,
-    unless: proc { |organisation| organisation.matched_effort_provider? },
+    if: proc { |organisation| organisation.is_reporter? },
     format: {with: /\A[a-zA-Z]{2,}-[a-zA-Z]{3}-.+\z/, message: I18n.t("activerecord.errors.models.organisation.attributes.iati_reference.format")}
   validates :beis_organisation_reference, uniqueness: {case_sensitive: false}
   validates :beis_organisation_reference,
     presence: true,
-    unless: proc { |organisation| organisation.matched_effort_provider? },
+    if: proc { |organisation| organisation.is_reporter? },
     format: {with: /\A[A-Z]{2,5}\z/, message: I18n.t("activerecord.errors.models.organisation.attributes.beis_organisation_reference.format")}
 
   scope :sorted_by_name, -> { order(name: :asc) }
   scope :delivery_partners, -> { sorted_by_name.where(role: "delivery_partner") }
+  scope :matched_effort_providers, -> { sorted_by_name.where(role: "matched_effort_provider") }
+  scope :active, -> { where(active: true) }
 
   before_validation :ensure_beis_organisation_reference_is_uppercase
 
@@ -41,6 +44,10 @@ class Organisation < ApplicationRecord
 
   def is_government?
     %w[10 11].include?(organisation_type)
+  end
+
+  def is_reporter?
+    service_owner? || delivery_partner?
   end
 
   def self.service_owner