+++ app/policies/matched_effort_policy.rb
@@ -1,0 +1,13 @@
+class MatchedEffortPolicy < ApplicationPolicy
+  def create?
+    Pundit.policy(user, record.activity).update?
+  end
+
+  def update?
+    create?
+  end
+
+  def destroy?
+    update?
+  end
+end