+++ config/routes.rb
@@ -23,7 +23,7 @@ Rails.application.routes.draw do
       end
     end
 
-    constraints role: /delivery_partners|matched_effort_providers/ do
+    constraints role: /delivery_partners|matched_effort_providers|external_income_providers/ do
       get "organisations/(:role)", to: "organisations#index", defaults: {role: "delivery_partners"}, as: :organisations
       get "organisations/(:role)/new", to: "organisations#new", as: :new_organisation
     end
@@ -36,6 +36,7 @@ Rails.application.routes.draw do
         resource :children, controller: :activity_children, only: [:show, :create]
 
         get "comments" => "activity_comments#show"
+        get "other_funding" => "activity_other_funding#show"
       end
     end
 
@@ -69,7 +70,11 @@ Rails.application.routes.draw do
       end
     end
 
-    resources :activities, only: [], concerns: [:transactionable, :budgetable, :disbursement_plannable] do
+    concern :matched_effortable do
+      resources :matched_efforts
+    end
+
+    resources :activities, only: [], concerns: [:transactionable, :budgetable, :disbursement_plannable, :matched_effortable] do
       resource :redaction, only: [:edit, :update], controller: :activity_redactions
       resources :steps, controller: "activity_forms"
       resources :implementing_organisations, only: [:new, :create, :edit, :update]