+++ spec/support/form_helpers.rb
@@ -204,6 +204,11 @@ module FormHelpers
     choose "GDI not applicable"
     click_button t("form.button.activity.submit")
 
+    expect(page).to have_content t("form.legend.activity.aid_type")
+    expect(page).to have_content t("form.hint.activity.aid_type")
+    choose("activity[aid_type]", option: aid_type)
+    click_button t("form.button.activity.submit")
+
     unless level == "fund"
       expect(page).to have_content t("form.label.activity.collaboration_type")
       choose "Bilateral"
@@ -225,11 +230,6 @@ module FormHelpers
       choose("activity[fund_pillar]", option: fund_pillar)
       click_button t("form.button.activity.submit")
     end
-
-    expect(page).to have_content t("form.legend.activity.aid_type")
-    expect(page).to have_content t("form.hint.activity.aid_type")
-    choose("activity[aid_type]", option: aid_type)
-    click_button t("form.button.activity.submit")
 
     if aid_type.in?(["B02", "B03", "D01"])
       expect(page).to have_content t("form.legend.activity.fstc_applies")
@@ -490,6 +490,27 @@ module FormHelpers
     fill_in "transfer[value]", with: transfer.value
 
     transfer
+  end
+
+  def fill_in_matched_effort_form(template = build(:matched_effort))
+    select template.organisation.name, from: "matched_effort[organisation_id]"
+
+    page.find(:xpath, "//input[@value='#{template.funding_type}']").set(true)
+    page.find(:xpath, "//input[@value='#{template.category}']").set(true)
+
+    fill_in "matched_effort[committed_amount]", with: template.committed_amount
+
+    within "#matched-effort-currency-field" do
+      find("option[value='#{template.currency}']").select_option
+    end
+
+    fill_in "matched_effort[exchange_rate]", with: template.exchange_rate
+    fill_in "matched_effort[date_of_exchange_rate(3i)]", with: template.date_of_exchange_rate.day
+    fill_in "matched_effort[date_of_exchange_rate(2i)]", with: template.date_of_exchange_rate.month
+    fill_in "matched_effort[date_of_exchange_rate(1i)]", with: template.date_of_exchange_rate.year
+    fill_in "matched_effort[notes]", with: template.notes
+
+    click_on t("default.button.submit")
   end
 
   private def activity_level(level)